\name{MVCM.binary-package}
\alias{MVCM.binary-package}
\alias{MVCM.binary}
\docType{package}
\title{
\packageTitle{MVCM.binary}
}
\description{
\packageDescription{MVCM.binary}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MVCM.binary}
\packageIndices{MVCM.binary}
main functions:
}
\author{
\packageAuthor{MVCM.binary}

Maintainer: \packageMaintainer{MVCM.binary}
}
\references{
TBD
}

\keyword{ package }
\seealso{

}
\examples{

}
