% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{SolveThetaA}
\alias{SolveThetaA}
\title{solve Theta and A}
\usage{
SolveThetaA(
  ThetaStart = NULL,
  AStart = NULL,
  Y,
  X,
  B,
  tolTheta = 1e-06,
  MaxItTheta = 50,
  lambda,
  gamma = 0,
  seed = 819,
  rank,
  tolAll = 1e-06,
  MaxItAll = 20,
  tolA = 1e-06,
  MaxItA = 50,
  tau = 1,
  m = 1,
  method = "LASSO",
  a = 3.7,
  ...
)
}
\arguments{
\item{ThetaStart}{Matrix. The startpoint of Theta, NULL is default.}

\item{Y}{A matrix.
A \eqn{q \times n} matrix, where the number of rows is the number of variables.}

\item{X}{A matrix. A \eqn{(p+1) \times n} matrix, where the number of rows is the number of variabels.}

\item{B}{A matrix. A \eqn{K \times r} matrix, where the \eqn{i}-th column is a plug-in spline basis at \eqn{T_i}.}

\item{tolTheta}{A number. \eqn{\epsilon_\Theta} is used in updating \eqn{\Theta} when \eqn{A} is fixed.}

\item{MaxItTheta}{A number. The maximum iteration for updating \eqn{\Theta} when \eqn{A} is fixed.}

\item{lambda}{A number. One of adaptive group lasso tuning parameters, \eqn{\lambda >0}.}

\item{gamma}{A number. One of adaptive group lasso tuning parameters, \eqn{\gamma=0} is equivalent to regular lasso.}

\item{seed}{A number. Seed value for randomness.}

\item{rank}{A number. The number of principal components, used to give a random \eqn{\Theta}.}

\item{tolAll}{A number. A measure of two \eqn{\Theta A^T}, it is used to determine the overall}

\item{MaxItAll}{A number. The maximum iteration for updating \eqn{\Theta A^T}.}

\item{tolA}{A number. Tolerance number for updating A.}

\item{MaxItA}{A number. The maximum iteration for updating \eqn{A}.}

\item{...}{Other parameters. Including control parameter in Stiefel Manifold.}
}
\value{
A list including \eqn{\Theta} and \eqn{A}.
}
\description{
solve Theta and A
}
