\name{MVCM.binary-package}
\alias{MVCM.binary-package}
\alias{MVCM.binary}
\docType{package}
\title{
\packageTitle{MVCM.binary}
}
\description{
\packageDescription{MVCM.binary}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MVCM.binary}
\packageIndices{MVCM.binary}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MVCM.binary}

Maintainer: \packageMaintainer{MVCM.binary}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
